Move exhaustive tests to common_tests
